refactor: rename compound and uniswap protocol files to v3
